refactor: change header and improve readability

+++ ft_lstsize_bonus.c
@@ -6,15 +6,16 @@
 /*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/12 17:53:50 by rde-mour          #+#    #+#             */
-/*   Updated: 2023/10/29 19:24:51 by rde-mour         ###   ########.org.br   */
+/*   Updated: 2024/09/27 11:50:56 by rde-mour         ###   ########.org.br   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stddef.h>
 
 int	ft_lstsize(t_list *list)
 {
-	int		size;
+	size_t	size;
 
 	if (!list)
 		return (0);
@@ -22,7 +23,7 @@ int	ft_lstsize(t_list *list)
 	while (list)
 	{
 		size++;
-		list = list -> next;
+		list = list->next;
 	}
 	return (size);
 }
